Migrate chenango (counting-group SOVC PDF, table header + text precincts)
